# openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Neuraal API",
    "version": "0.1.0",
    "description": "Backend API for Neuraal — task/note management with AI features (summaries, auto-topic classification)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "DevUserIdHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "x-user-id",
        "description": "Temporary dev-only auth. Sends a fixed user ID. Will be replaced by BearerAuth (JWT) in production."
      },
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Future JWT auth (not enforced yet)."
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "object",
            "required": [
              "code",
              "message"
            ],
            "properties": {
              "code": {
                "type": "string",
                "example": "NOT_FOUND"
              },
              "message": {
                "type": "string",
                "example": "Resource not found"
              }
            }
          }
        }
      },
...
      "Entry": {
        "type": "object",
        "required": [
          "id",
          "userId",
          "date",
          "type",
          "title",
          "content",
          "version",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string"
          },
          "date": {
            "type": "string",
            "pattern": "^\\d{4}-\\d{2}-\\d{2}$",
            "example": "2025-06-15"
          },
          "type": {
            "type": "string",
            "enum": [
              "task",
              "note"
            ]
          },
          "title": {
            "type": "string",
            "maxLength": 120
          },
          "content": {
            "type": "object",
            "additionalProperties": true,
            "description": "TipTap/ProseMirror JSON content"
          },
          "topicId": {
            "type": [
              "string",
              "null"
            ]
          },
          "completed": {
            "type": [
              "boolean",
              "null"
            ],
            "description": "null for notes"
          },
          "version": {
            "type": "integer",
            "minimum": 1
          },
          "sortOrder": {
            "type": "integer",
            "minimum": 0,
            "description": "Display order within a day. Lower values appear first."
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "summary": {
            "type": [
              "string",
              "null"
            ]
          },
          "summaryFormat": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "markdown",
              "plain",
              null
            ]
          },
          "summaryUpdatedAt": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    },
    "responses": {
      "BadRequest": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Authentication required",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "NotFound": {
        "description": "Resource not found",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "Conflict": {
        "description": "Conflict (duplicate or version mismatch)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      },
      "InternalError": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorResponse"
            }
          }
        }
      }
    },
    "parameters": {
      "ResourceId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string",
          "format": "uuid"
        },
        "description": "Resource UUID"
      }
    }
  },
  "security": [
    {
      "DevUserIdHeader": []
    }
  ],
  "paths": {
...
    "/api/entries": {
      "get": {
        "tags": [
          "Entries"
        ],
        "summary": "List entries by date",
        "operationId": "listEntries",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
            },
            "description": "Date in YYYY-MM-DD format"
          }
        ],
        "responses": {
          "200": {
            "description": "List of entries for the given date",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "entries"
                  ],
                  "properties": {
                    "entries": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Entry"
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      },
      "post": {
        "tags": [
          "Entries"
        ],
        "summary": "Create an entry",
        "operationId": "createEntry",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "date",
                  "type",
                  "title",
                  "content"
                ],
                "properties": {
                  "date": {
                    "type": "string",
                    "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "task",
                      "note"
                    ]
                  },
                  "title": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "content": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "topicId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "completed": {
                    "type": "boolean"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Entry created",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "entry"
                  ],
                  "properties": {
                    "entry": {
                      "$ref": "#/components/schemas/Entry"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/api/entries/{id}": {
      "patch": {
        "tags": [
          "Entries"
        ],
        "summary": "Update an entry",
        "operationId": "updateEntry",
        "description": "Updates entry fields with optimistic concurrency (version check).",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "version": {
                    "type": "integer",
                    "description": "Current version for optimistic concurrency"
                  },
                  "title": {
                    "type": "string",
                    "maxLength": 120
                  },
                  "content": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "topicId": {
                    "type": [
                      "string",
                      "null"
                    ]
                  },
                  "completed": {
                    "type": "boolean"
                  },
                  "type": {
                    "type": "string",
                    "enum": [
                      "task",
                      "note"
                    ],
                    "description": "Change entry type"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Entry updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "entry"
                  ],
                  "properties": {
                    "entry": {
                      "$ref": "#/components/schemas/Entry"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      },
      "delete": {
        "tags": [
          "Entries"
        ],
        "summary": "Delete an entry",
        "operationId": "deleteEntry",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "204": {
            "description": "Entry deleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          }
        }
      }
    },
...
    "/api/entries/{id}/summarize": {
      "post": {
        "tags": [
          "Entries"
        ],
        "summary": "Request entry summary (async)",
        "operationId": "requestEntrySummary",
        "description": "Enqueues an async AI summary generation. Returns 202 Accepted.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "responses": {
          "202": {
            "description": "Summary generation started",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "requestId",
                    "notificationId",
                    "message"
                  ],
                  "properties": {
                    "requestId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "notificationId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          }
        }
      }
    },
    "/api/entries/{id}/auto-topic": {
      "post": {
        "tags": [
          "Embeddings"
        ],
        "summary": "Auto-assign topic to entry",
        "operationId": "autoAssignTopic",
        "description": "Generates an embedding for the entry text, finds the best matching topic by cosine similarity, and assigns it if the score meets the threshold.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ResourceId"
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "threshold": {
                    "type": "number",
                    "minimum": 0,
                    "maximum": 1,
                    "description": "Similarity threshold (default 0.35). Higher = stricter."
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Auto-topic result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "entryId",
                    "selectedTopicId",
                    "score"
                  ],
                  "properties": {
                    "entryId": {
                      "type": "string",
                      "format": "uuid"
                    },
                    "selectedTopicId": {
                      "type": [
                        "string",
                        "null"
                      ],
                      "format": "uuid"
                    },
                    "score": {
                      "type": [
                        "number",
                        "null"
                      ],
                      "description": "Similarity score 0..1"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "500": {
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  }
}